Updates node version from 6.12 to 10.16
As of 30 April 2019, NodeJS 6.0 is no longer supported (https://nodejs.org/en/blog/release/v6.17.1/)
This commit updates the engine to the latest lts version which is 10.16.0

It also introduces a `.nvmrc` file to developers who might have to work
on or support services using different versions of node. This is only optional
so shouldn't be an issue.